helm-version-upgrade/pyproject.toml
[tool.poetry]
name = "helm-chart-version-upgrade"
version = "1.0.0"
description = "A GitHub Action to upgrade a helm chart version"
authors = ["Pascal Holthaus <pascal.holthaus@greenbone.net>"]
license = "GPL-3"
packages = [
    { include = "action" },
]

[tool.poetry.dependencies]
python = "^3.9"
ruamel-yaml = "^0.17.21"

[tool.poetry.dev-dependencies]
autohooks-plugin-pylint = ">=22.8.0"
autohooks-plugin-isort = ">=22.8.0"
autohooks-plugin-black = ">=22.8.0"

[tool.poetry.scripts]
helm-chart-version-upgrade = 'action.helm_chart_version_upgrade:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py39', 'py310', 'py311']
exclude = '''
/(
    \.git
  | \.venv
  | \.github
  | \.vscode
  | _build
  | build
  | dist
  | docs
)/
'''

[tool.isort]
profile = "black"
line_length = 80

[tool.autohooks]
mode = "poetry"
pre-commit = ['autohooks.plugins.black', 'autohooks.plugins.isort', 'autohooks.plugins.pylint']
